+++ StabilityMatrix.Avalonia/ViewModels/Inference/InferenceImageToVideoViewModel.cs
@@ -199,6 +199,7 @@ public partial class InferenceImageToVideoViewModel
                 OutputNodeNames = buildPromptArgs.Builder.Connections.OutputNodeNames.ToArray(),
                 Parameters = SaveStateToParameters(new GenerationParameters()),
                 Project = InferenceProjectDocument.FromLoadable(this),
+                FilesToTransfer = buildPromptArgs.FilesToTransfer,
                 // Only clear output images on the first batch
                 ClearOutputImages = i == 0
             };