+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to Stability Matrix will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning 2.0](https://semver.org/spec/v2.0.0.html).
 
+## v2.9.0-dev.2
+### Added
+#### Inference
+- Added option to load a .yaml config file next to the model with the same name. Can be used with VPred and other models that require a config file.
+
 ## v2.9.0-dev.1
 ### Added
 - Added new package: [StableSwarmUI](https://github.com/Stability-AI/StableSwarmUI) by Stability AI
@@ -16,6 +21,7 @@ and this project adheres to [Semantic Versioning 2.0](https://semver.org/spec/v2
 ### Fixed
 - Fixed user tokens read error causing failed downloads
 - Failed downloads will now log error messages
+- Fixed [#458](https://github.com/LykosAI/StabilityMatrix/issues/458) - Save Intermediate Image not working
 
 ## v2.8.2
 ### Added